refactor: don't log lvol ptpl destroy errors when not found
If the lvol was never shared, then the ptpl will not exist.

+++ io-engine/src/lvs/lvs_lvol.rs
@@ -571,7 +571,9 @@ impl PtplFileOps for LvolPtpl {
 
     fn destroy(&self) -> Result<(), std::io::Error> {
         if let Some(path) = self.path() {
-            std::fs::remove_file(path)?;
+            if path.exists() {
+                std::fs::remove_file(path)?;
+            }
         }
         Ok(())
     }
